authentication bases
