+++ app/controllers/api_controller.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class APIController < ApplicationController
+  include SessionManagement
+end
